+++ CHANGELOG.md
@@ -20,10 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initialize services on startup before app load
 - Improved wording on several error dialogs
 - Flower worker is now more responsive (100ms -> 16ms delay)
+- Client versions (`Versions.xml`) are only saved when the existing file does not exist
+- Color themes (`Themes.xml`) are only saved when the existing file does not exist
 
 ### Removed
 
-- Annoying "unable to save character state" popups on close, are logged instead
+- Annoying "unable to save file" popups on close, are logged instead
 - More dead code
 
 ## [4.1.0] - 2023-04-10